requirements.txt
fastapi 
uvicorn
scrapegraph-py
pydantic
sqlalchemy
google-search-results
nest_asyncio
python-dotenv